properly respect executable bit
bashes `source` doesn't care if the executable bit is set on a file
it just sources it verbatim. this was not the most ideal, so we've
now required the executable bit to be set in order to source a file.
(sash_add has also been changed to run: `chmod +x` on the file) so
there's zero service interruption.